+++ villager-tweaks/common/src/main/java/com/chimericdream/villagertweaks/entity/VT$VillagerAccessor.java
@@ -1,0 +1,9 @@
+package com.chimericdream.villagertweaks.entity;
+
+import net.minecraft.world.entity.npc.villager.VillagerData;
+
+public interface VT$VillagerAccessor {
+    default VillagerData getVillagerData() {
+        return null;
+    }
+}