+++ villager-tweaks/CHANGELOG.md
@@ -1,4 +1,4 @@
-### Unreleased changes
+### 26.2 - 6.2.0
 
 #### New Features